refactor: rename data class
