+++ src/modules/todo/components/item.vue
@@ -20,7 +20,7 @@
 
 <script lang='ts'>
 import { Component, Vue, Prop } from 'nuxt-property-decorator';
-import { Todo } from '../data/todo';
+import Todo from '../store/state/todo';
 import { Actions } from '../../../mixins/store_helper';
 import { TodoActions } from '../store/actions';
 import { HTMLElementEvent } from '../../../types';
