+++ src/modules/todo/store/actions.ts
@@ -1,5 +1,5 @@
 import { DefineActions } from 'vuex-type-helper';
-import { Todo } from '../data/todo';
+import Todo from './state/todo';
 import { TodoState } from './state';
 import { TodoMutations } from './mutations';
