+++ src/modules/todo/store/getters.ts
@@ -1,5 +1,5 @@
 import { DefineGetters } from 'vuex-type-helper';
-import { Todo } from '../data/todo';
+import Todo from './state/todo';
 import { TodoState } from './state';
 
 export interface TodoGetters {
